适配 BaseLib 3.1.6，修复进阶之蛇描述与蛇之心计数
BaseLib 3.1.6 适配:
- JormungandrAncient: OptionPools.Roll(Rng) -> Roll(Rng, this) 修复 MissingMethodException
- FrozenBloodRelic: SavedProperty decimal -> int，避免保存时 JsonException
- RecordPower: SavedProperty decimal -> int，避免保存时 JsonException
- SnakeVenomTherapyPendingPower: SavedProperty decimal -> int
- SnakeVenomTherapyCard: PoisonAmount 赋值增加 (int) 转换

进阶之蛇相关:
- AdvancedSnakePatches: 新增 DynamicDescription Postfix 动态注入描述
- AdvancedSnakePatches: 新增 EnchantmentModel.CanEnchant Prefix 禁止进阶之福被附魔
- AdvancedSnakeRelic: BuildEffectsDescription 改为 internal 供补丁访问
- relics.json: 微调蛇之心描述文本

蛇之心计数修复:
- 获得遗物后计数初始值 1 -> 0
- 触发复活后计数重置 1 -> 0
- 计数循环现为 0 -> 1 -> 2 -> 3(Active) -> 触发 -> 0

其他:
- .gitignore: BaseLib 目录 3.0.9 -> 3.1.6

+++ slay-the-spire-2/Scripts/cards/SnakeVenomTherapyCard.cs
@@ -41,7 +41,7 @@ public class SnakeVenomTherapyCard : SnakeTheBiteCardModel
     {
         await CreatureCmd.TriggerAnim(Owner.Creature, "Cast", Owner.Character.CastAnimDelay);
         await CreatureCmd.Heal(Owner.Creature, DynamicVars.Heal.BaseValue);
-        var power = new SnakeVenomTherapyPendingPower { SnakeTheBite_PoisonAmount = DynamicVars.Poison.BaseValue };
+        var power = new SnakeVenomTherapyPendingPower { SnakeTheBite_PoisonAmount = (int)DynamicVars.Poison.BaseValue };
         await PowerCmd.Apply(power, Owner.Creature, 2m, Owner.Creature, this);
     }
