+++ slay-the-spire-2/Scripts/powers/SnakeVenomTherapyPendingPower.cs
@@ -17,7 +17,7 @@ public class SnakeVenomTherapyPendingPower : SnakeTheBitePowerModel
     public override PowerStackType StackType => PowerStackType.Counter;
     public override int DisplayAmount => (int)Amount;
 
-    [SavedProperty] public decimal SnakeTheBite_PoisonAmount { get; set; }
+    [SavedProperty] public int SnakeTheBite_PoisonAmount { get; set; }
 
     // 玩家回合结束时递减计数器，到1时触发中毒并移除
     public override async Task BeforeTurnEnd(PlayerChoiceContext choiceContext, CombatSide side)
